test: 검증 로직을 MembershipValidatorTest로 이동
